FLUORESCENCIA/requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.0
asgiref==3.2.10
beautifulsoup4==4.9.0
cli-helpers==2.1.0
click==7.1.2
configobj==5.0.6
crispy-forms-gds==0.2.1
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.2
django-bootstrap4==1.1.1
django-cleanup==5.1.0
django-crispy-forms==1.9.0
django-static-fontawesome==5.14.0.0
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
humanize==3.1.0
imageio==2.8.0
install==1.3.4
Jinja2==2.11.2
joblib==0.15.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.0
mpld3==0.3
mpmath==1.1.0
networkx==2.5
numpy==1.18.1
opencv-python==4.4.0.44
pandas==1.0.4
pgspecial==1.11.10
Pillow==7.1.2
prompt-toolkit==3.0.8
psycopg2==2.8.4
psycopg2-binary==2.8.6
Pygments==2.7.2
pyparsing==2.4.6
pyPdf==1.13
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
reportlab==3.5.55
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.3
six==1.15.0
soupsieve==2.0
sqlparse==0.3.1
tabulate==0.8.7
terminaltables==3.1.0
threadpoolctl==2.1.0
tifffile==2020.10.1
Unipath==1.1
virtualenv==20.1.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
